Strike vs decay plot, corrected num_strikes column in df_bids, added list of distinct strikes in df_bids
